# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  /** @nullable */
  rateLimitReset?: number | null;
}

export interface GithubUser {
  login: string;
  /** @nullable */
  name?: string | null;
  avatar_url: string;
  /** @nullable */
  bio?: string | null;
  followers: number;
  following: number;
  public_repos: number;
  html_url: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  blog?: string | null;
  created_at: string;
  cached: boolean;
}

export interface GithubRepo {
  id: number;
  name: string;
  full_name: string;
  /** @nullable */
  description?: string | null;
  html_url: string;
  /** @nullable */
  language?: string | null;
  stargazers_count: number;
  forks_count: number;
  open_issues_count: number;
  default_branch: string;
  updated_at: string;
  created_at: string;
  fork: boolean;
  topics: string[];
  visibility?: string;
  watchers_count: number;
}

export interface RepoListResponse {
  repos: GithubRepo[];
  total_count: number;
  page: number;
  per_page: number;
  has_more: boolean;
  cached: boolean;
}

export interface LanguageEntry {
  language: string;
  count: number;
}

export interface LanguageStats {
  languages: LanguageEntry[];
  cached: boolean;
}

export type GetGithubUserParams = {
username: string;
};

export type GetGithubReposParams = {
username: string;
sort?: GetGithubReposSort;
page?: number;
per_page?: number;
};

export type GetGithubReposSort = typeof GetGithubReposSort[keyof typeof GetGithubReposSort];


export const GetGithubReposSort = {
  stars: 'stars',
  name: 'name',
  updated: 'updated',
} as const;

export type GetGithubLanguagesParams = {
username: string;
};