# lib/api-zod/src/generated/types/githubRepo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface GithubRepo {
  id: number;
  name: string;
  full_name: string;
  /** @nullable */
  description?: string | null;
  html_url: string;
  /** @nullable */
  language?: string | null;
  stargazers_count: number;
  forks_count: number;
  open_issues_count: number;
  default_branch: string;
  updated_at: string;
  created_at: string;
  fork: boolean;
  topics: string[];
  visibility?: string;
  watchers_count: number;
}