# lib/api-zod/src/generated/types/githubUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface GithubUser {
  login: string;
  /** @nullable */
  name?: string | null;
  avatar_url: string;
  /** @nullable */
  bio?: string | null;
  followers: number;
  following: number;
  public_repos: number;
  html_url: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  blog?: string | null;
  created_at: string;
  cached: boolean;
}