# lib/api-zod/src/generated/types/languageEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface LanguageEntry {
  language: string;
  count: number;
}