# lib/api-zod/src/generated/types/languageStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { LanguageEntry } from './languageEntry';

export interface LanguageStats {
  languages: LanguageEntry[];
  cached: boolean;
}