# lib/api-zod/src/generated/types/repoListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Repo Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { GithubRepo } from './githubRepo';

export interface RepoListResponse {
  repos: GithubRepo[];
  total_count: number;
  page: number;
  per_page: number;
  has_more: boolean;
  cached: boolean;
}